Move FileUtils to common module so they can be accessed anywhere. Remove interface FileUtils and rename implementation to FileUtils
