+++ orchestrator/src/test/java/com/scottlogic/deg/orchestrator/violate/ViolateOutputValidatorTests.java
@@ -18,7 +18,7 @@ package com.scottlogic.deg.orchestrator.violate;
 
 import com.scottlogic.deg.common.profile.Profile;
 import com.scottlogic.deg.common.profile.Rule;
-import com.scottlogic.deg.output.FileUtils;
+import com.scottlogic.deg.common.util.FileUtils;
 import com.scottlogic.deg.output.OutputPath;
 import com.scottlogic.deg.output.outputtarget.OutputTargetValidationException;
 import org.junit.Test;
