+++ output/src/main/java/com/scottlogic/deg/output/outputtarget/FileOutputTarget.java
@@ -19,7 +19,7 @@ package com.scottlogic.deg.output.outputtarget;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
 import com.scottlogic.deg.common.profile.ProfileFields;
-import com.scottlogic.deg.output.FileUtils;
+import com.scottlogic.deg.common.util.FileUtils;
 import com.scottlogic.deg.output.OutputPath;
 import com.scottlogic.deg.output.writer.DataSetWriter;
 import com.scottlogic.deg.output.writer.OutputWriterFactory;
@@ -28,7 +28,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 
 public class FileOutputTarget implements SingleDatasetOutputTarget {
     private final Path filePath;
